+++ lib/icon_content.dart
@@ -1,9 +1,5 @@
 import 'package:flutter/material.dart';
-
-const labelTextStyle = TextStyle(
-  fontSize: 18.0,
-  color: Color(0xFF8D8E98),
-);
+import 'constants.dart';
 
 // カードの中身（性別選択）
 class IconContent extends StatelessWidget {
